[BE] SDK Enhancement Documentation and Add SDK Identifier 2 - Go
[BE] SDK Enhancement Documentation and Add SDK Identifier 2 - Go
