+++ models/va/updateVa/updateVaAdditionalInfoDTO.go
@@ -1,6 +1,11 @@
 package models
 
+import (
+	createVaModels "github.com/PTNUSASATUINTIARTHA-DOKU/doku-golang-library/models/va/createVa"
+)
+
 type UpdateVaAdditionalInfoDTO struct {
 	Channel              string                          `json:"channel"`
 	VirtualAccountConfig UpdateVaVirtualAccountConfigDTO `json:"virtualAccountConfig"`
+	Origin               createVaModels.Origin           `json:"origin"`
 }
